lsa: fix printing object_descriptor::_alignment
object_descriptor::_alignment is of type uint8_t which is actually an
unsigned char.

+++ utils/logalloc.cc
@@ -656,7 +656,7 @@ class region_impl : public allocation_strategy {
 
         friend std::ostream& operator<<(std::ostream& out, const object_descriptor& desc) {
             return out << sprint("{flags = %x, migrator=%p, alignment=%d, size=%d}",
-                (int)desc._flags._value, desc._migrator, desc._alignment, desc._size);
+                (int)desc._flags._value, desc._migrator, unsigned(desc._alignment), desc._size);
         }
     } __attribute__((packed));
 private:
